Upgrade GitHub Actions, mypy and pre-commit hooks (#13)
* Update GitHub Actions to latest version

This also pins us to the major version, only, so we are using the same version
across all workflows. Otherwise it doesn't seem like Dependabot is able to
update us with the latest versions.

* Update mypy to version 0.971

This is the latest version available.

* Upgrade pre-commit hooks to the latest version

* Rename "test" dependency group to "dev"

This is the default name of dev-only dependencies. I'm not sure I got it wrong
with "test" when I changed the file to Poetry 1.2 format.