Print the difference to an old report to STDERR
By comparing against an old report, we can find and print the new errors
that are found to STDERR. This means we don't have to re-run Mypy to get
the line number and hints for fixing the error.

Because we don't store the line number in the reports, we can't be sure
which instance of an error in a file is the new one, so we print all
instances of an error in a file. For good measure, we also print all
other errors on those lines.

Note: we can't pipe the output of this command to the file that we're
reading from, because by the time we get to read it, it will have
already been overwritten. As such, if we wish to store the JSON report
from a run that takes the `--diff-old-report/-d` flag into the same
file, then the `--output-file/-o` flag must be used to capture the
output.