Fix the release workflow
It actually didn't break when I tried to release version 1.0.3, but that was
pure luck because the `pypa/gh-action-pypi-publish` GitHub Action took a while
to get set up, so it didn't end up trying to fetch the release files before
they had been uploaded.

Specifying the dependency however makes sure we won't have a problem with this
going forward.