Change spacing around report summary
Before this change an empty line would be printed above the summary when
there were no new errors, and no empty line would be printed when there
were errors.

This change reverses that. An empty line is printed to separate the
summary from the new errors only if there are new errors.