Handle errors that don't mention a line number
This is done by stopping the parsing entirely and letting the user know why so
they can investigate and hopefully fix the error first.

Errors that mention a filename but no specific line are normally indicators of
a big issue which we don't have a strategy for handling right now.